+++ tests/test_security.py
@@ -1,8 +1,15 @@
+from unittest.mock import patch
+
 import pytest
 from fastapi import HTTPException
 from fastapi.security import HTTPBasicCredentials
+from ldap3.core.exceptions import LDAPBindError
 
-from mex.backend.security import has_read_access, has_write_access
+from mex.backend.security import (
+    has_read_access,
+    has_write_access,
+    has_write_access_ldap,
+)
 
 read_credentials = HTTPBasicCredentials(
     username="Reader",
@@ -96,3 +103,30 @@ def test_has_read_access_with_basic_auth() -> None:
     with pytest.raises(HTTPException) as error:
         has_read_access(credentials=None)
     assert "Missing authentication" in error.value.detail
+
+
+def test_has_write_access_ldap_success() -> None:
+    username = "testuser"
+    password = "right_password"  # noqa: S105
+    with patch("mex.backend.security.BackendSettings.get") as mock_settings:
+        mock_settings.return_value.ldap_url.get_secret_value.return_value = (
+            "ldaps://ldap.example.com:636"
+        )
+        with patch("mex.backend.security.Connection") as mock_connection:
+            mocked_connection = mock_connection.return_value.__enter__.return_value
+            mocked_connection.server.check_availability.return_value = True
+            assert has_write_access_ldap(username, password) is True
+
+
+def test_has_write_access_ldap_bind_error() -> None:
+    username = "testuser"
+    password = "wrong_password"  # noqa: S105
+    with patch("mex.backend.security.BackendSettings.get") as mock_settings:
+        mock_settings.return_value.ldap_url.get_secret_value.return_value = (
+            "ldaps://ldap.example.com:636"
+        )
+        # Patch Connection to raise LDAPBindError when instantiated
+        with patch(
+            "mex.backend.security.Connection", side_effect=LDAPBindError("fail")
+        ):
+            assert has_write_access_ldap(username, password) is False
